+++ test/resources.test.ts
@@ -1,44 +1,17 @@
-import { gn, rdf, schema, sh } from '@tpluscode/rdf-ns-builders/strict'
+import { gn, rdf, schema, sh } from '@tpluscode/rdf-ns-builders'
 import { expect } from 'chai'
 import $rdf from 'rdf-ext'
 import clownface from 'clownface'
-import { meta } from '../lib/ns.js'
+import { meta } from '@zazuko/vocabulary-extras/builders'
 import { children, example } from '../resources.js'
 import { client, streamClient } from './client.js'
-import { ex, parse, startFuseki, testData } from './support.js'
+import { ex, parse, startFuseki } from './support.js'
+import { insertGeoData } from './testData.js'
 
 describe('@zazuko/cube-hierarchy-query/resources', () => {
   before(startFuseki)
 
-  before(async function () {
-    await testData`
-      <US> a ${ex.Country} ; ${schema.containedInPlace} <North-America> .
-      <BR> a ${ex.Country} ; ${schema.containedInPlace} <South-America> .
-      <AR> a ${ex.Country} ; ${schema.containedInPlace} <South-America> .
-      <VE> a ${ex.Country} ; ${schema.containedInPlace} <South-America> .
-
-      <CH> a ${ex.Country} ; ${schema.containedInPlace} <Europe> .
-
-      <https://sws.geonames.org/2658434>
-        a ${gn.A} ; ${gn.parentFeature} <Europe> .
-
-
-      <https://sws.geonames.org/798544>
-        a ${gn.A} ; ${schema.containedInPlace} <Europe> .
-
-      <ZH> a ${ex.Canton} ; ${schema.containedInPlace} <CH> .
-
-      <Affoltern> a ${ex.District} ;
-        ${schema.containedInPlace} <ZH> ;
-        ${schema.containsPlace} <Bonstetten>, <Rifferswil>, <Stallikon>, <Knonau> ;
-      .
-
-      <Bonstetten> a ${ex.Municipality} ; ${schema.name} "Bonstetten" .
-      <Rifferswil> a ${ex.Municipality} ; ${schema.name} "Rifferswil" .
-      <Stallikon> a ${ex.Municipality} ; ${schema.name} "Stallikon" .
-      <Knonau> a ${ex.Municipality} ; ${schema.name} "Knonau" .
-    `
-  })
+  before(insertGeoData)
 
   describe('example', () => {
     it('returns empty query when last property id not defined', async () => {
@@ -121,7 +94,7 @@ describe('@zazuko/cube-hierarchy-query/resources', () => {
       const query = children(hierarchy.namedNode(ex.firstLevel), ex('South-America'), {
         limit: 1,
       })
-      const result = await query.execute(streamClient, $rdf)
+      const { children: result } = await query.execute(streamClient, $rdf)
 
       // then
       expect(result).to.have.length(1)
@@ -163,7 +136,7 @@ describe('@zazuko/cube-hierarchy-query/resources', () => {
       const query = children(hierarchy.namedNode(ex.districtLevel), ex('Affoltern'), {
         limit: 1,
       })
-      const result = await query.execute(streamClient, $rdf)
+      const { children: result } = await query.execute(streamClient, $rdf)
 
       // then
       expect(result).to.have.length(1)
@@ -207,7 +180,7 @@ describe('@zazuko/cube-hierarchy-query/resources', () => {
         offset: 2,
         orderBy: [schema.position, schema.name],
       })
-      const [result] = await query.execute(streamClient, $rdf)
+      const { children: [result] } = await query.execute(streamClient, $rdf)
 
       // then
       expect(result.term).to.deep.eq(ex.Rifferswil)